+++ fabric/src/main/java/rbasamoyai/createbigcannons/fabric/mixin/CannonMountBlockEntityMixin.java
@@ -1,0 +1,37 @@
+package rbasamoyai.createbigcannons.fabric.mixin;
+
+import javax.annotation.Nullable;
+
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
+
+import com.simibubi.create.foundation.blockEntity.SmartBlockEntity;
+
+import net.fabricmc.fabric.api.transfer.v1.item.ItemVariant;
+import net.fabricmc.fabric.api.transfer.v1.storage.Storage;
+import net.fabricmc.fabric.api.transfer.v1.storage.base.SidedStorageBlockEntity;
+import net.minecraft.core.BlockPos;
+import net.minecraft.core.Direction;
+import net.minecraft.world.level.block.entity.BlockEntityType;
+import net.minecraft.world.level.block.state.BlockState;
+import rbasamoyai.createbigcannons.cannon_control.cannon_mount.CannonMountBlockEntity;
+import rbasamoyai.createbigcannons.cannon_control.contraption.PitchOrientedContraptionEntity;
+
+@Mixin(CannonMountBlockEntity.class)
+public abstract class CannonMountBlockEntityMixin extends SmartBlockEntity implements SidedStorageBlockEntity {
+
+	CannonMountBlockEntityMixin(BlockEntityType<?> typeIn, BlockPos pos, BlockState state) {
+		super(typeIn, pos, state);
+	}
+
+	@Shadow
+	protected PitchOrientedContraptionEntity mountedContraption;
+
+	@Nullable
+	@Override
+	public Storage<ItemVariant> getItemStorage(@Nullable Direction face) {
+		return this.mountedContraption instanceof SidedStorageBlockEntity transferable ? transferable.getItemStorage(face) : null;
+	}
+
+
+}