+++ fabric/src/main/java/rbasamoyai/createbigcannons/fabric/mixin/YawControllerBlockEntityMixin.java
@@ -1,0 +1,32 @@
+package rbasamoyai.createbigcannons.fabric.mixin;
+
+import javax.annotation.Nullable;
+
+import org.spongepowered.asm.mixin.Mixin;
+
+import com.simibubi.create.content.kinetics.base.KineticBlockEntity;
+
+import net.fabricmc.fabric.api.transfer.v1.item.ItemVariant;
+import net.fabricmc.fabric.api.transfer.v1.storage.Storage;
+import net.fabricmc.fabric.api.transfer.v1.storage.base.SidedStorageBlockEntity;
+import net.minecraft.core.BlockPos;
+import net.minecraft.core.Direction;
+import net.minecraft.world.level.block.entity.BlockEntityType;
+import net.minecraft.world.level.block.state.BlockState;
+import rbasamoyai.createbigcannons.cannon_control.cannon_mount.CannonMountBlockEntity;
+import rbasamoyai.createbigcannons.cannon_control.cannon_mount.ExtendsCannonMount;
+import rbasamoyai.createbigcannons.cannon_control.cannon_mount.YawControllerBlockEntity;
+
+@Mixin(YawControllerBlockEntity.class)
+public abstract class YawControllerBlockEntityMixin extends KineticBlockEntity implements ExtendsCannonMount, SidedStorageBlockEntity {
+
+	YawControllerBlockEntityMixin(BlockEntityType<?> typeIn, BlockPos pos, BlockState state) { super(typeIn, pos, state); }
+
+	@Nullable
+	@Override
+	public Storage<ItemVariant> getItemStorage(@Nullable Direction face) {
+		CannonMountBlockEntity mount = this.getCannonMount();
+		return mount instanceof SidedStorageBlockEntity transferable ? transferable.getItemStorage(face) : null;
+	}
+
+}