+++ forge/src/main/java/rbasamoyai/createbigcannons/forge/mixin/YawControllerBlockEntityMixin.java
@@ -1,0 +1,34 @@
+package rbasamoyai.createbigcannons.forge.mixin;
+
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+
+import org.spongepowered.asm.mixin.Mixin;
+
+import com.simibubi.create.content.kinetics.base.KineticBlockEntity;
+
+import net.minecraft.core.BlockPos;
+import net.minecraft.core.Direction;
+import net.minecraft.world.level.block.entity.BlockEntityType;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraftforge.common.capabilities.Capability;
+import net.minecraftforge.common.util.LazyOptional;
+import rbasamoyai.createbigcannons.cannon_control.cannon_mount.CannonMountBlockEntity;
+import rbasamoyai.createbigcannons.cannon_control.cannon_mount.ExtendsCannonMount;
+import rbasamoyai.createbigcannons.cannon_control.cannon_mount.YawControllerBlockEntity;
+
+@Mixin(YawControllerBlockEntity.class)
+public abstract class YawControllerBlockEntityMixin extends KineticBlockEntity implements ExtendsCannonMount {
+
+	YawControllerBlockEntityMixin(BlockEntityType<?> typeIn, BlockPos pos, BlockState state) {
+		super(typeIn, pos, state);
+	}
+
+	@Nonnull
+	@Override
+	public <T> LazyOptional<T> getCapability(@Nonnull Capability<T> cap, @Nullable Direction side) {
+		CannonMountBlockEntity cannonMount = this.getCannonMount();
+		return cannonMount == null ? super.getCapability(cap, side) : cannonMount.getCapability(cap, side);
+	}
+
+}